Restructure event listeners
Events listeners are now opened in context manager blocks. This lets you
set up the listener _before_ sending the command that will trigger the
event. The old way was to set up the listener after sending the command,
and this led to very reproducible race conditions where the listener is
set up _after_ the event has already fired. This new approach is much
more robust.