Cancel nursery before closing connection (fixes #5)
This regression was introduced at a8cbe89 because as I refactored I
mistakenly removed the nursery.cancel_scope.cancel(), which is still
needed. This isn't the cleanest way to handle connection closed, but
it's the smallest change that fixes the build, and its based on code
that existed in the last successful build.