Refactor handling of closed WebSocket connection (fixes #5)
This is refactored to be more similar to the way trio-websocket works:
the reader task should will catch ConnectionClosed and quietly exit. The
public APIs (i.e. session.execute()) will raise ConnectionClosed
instead. It's easier for the caller to catch an exception in a
foreground task than in a background task.