Simplified API
This is inpsired by PR #6 but with a couple changes:

1. Use context vars instead of a singleton instance.
2. Use code generation instead of decoration. This improves type
   inference for tools like MyPy.
3. Extend the same changes to the connection class.

This commit includes the generator, changes to the trio_cdp module to
support the simplified API, and updated examples--including an example
of running inside Jupyter notebook.

The generated code will be added in the next commit to make review
easier.

Co-authored-by: Brian Mackintosh <bckmackintosh@gmail.com>